Moved db, config, cli functions to lib directory
