+++ bin/lib/cli/index.ts
