+++ bin/lib/config/index.ts
