+++ bin/lib/config/lcs.config.ts
