+++ bin/lib/config/lcs.pathes.ts
